Move `verifySourceRootsIntersectionsBasedOnAndroidVariants` to `DokkaGenerateTask`
