+++ dokka-runners/dokka-gradle-plugin/src/main/kotlin/engine/parameters/builders/DokkaParametersBuilder.kt
@@ -13,7 +13,6 @@ import org.jetbrains.dokka.*
 import org.jetbrains.dokka.gradle.DokkaBasePlugin
 import org.jetbrains.dokka.gradle.engine.parameters.DokkaGeneratorParametersSpec
 import org.jetbrains.dokka.gradle.engine.parameters.DokkaModuleDescriptionKxs
-import org.jetbrains.dokka.gradle.engine.parameters.DokkaSourceSetSpec
 import org.jetbrains.dokka.gradle.engine.plugins.DokkaPluginParametersBaseSpec
 import org.jetbrains.dokka.gradle.formats.DokkaHtmlPlugin.Companion.extractDokkaPluginMarkers
 import org.jetbrains.dokka.gradle.internal.InternalDokkaGradlePluginApi
@@ -39,10 +38,7 @@ internal class DokkaParametersBuilder(
         outputDirectory: File,
         cacheDirectory: File? = null,
         moduleDescriptorDirs: Iterable<File>,
-        projectDir: File
     ): DokkaConfiguration {
-        verifySourceRootsIntersectionsBasedOnAndroidVariants(spec.dokkaSourceSets, projectDir)
-
         val moduleName = spec.moduleName.get()
         val moduleVersion = spec.moduleVersion.orNull?.takeIf { it != Project.DEFAULT_VERSION }
         val offlineMode = spec.offlineMode.get()
@@ -154,104 +150,5 @@ internal class DokkaParametersBuilder(
             serializationFormat = DokkaConfiguration.SerializationFormat.JSON,
             values = spec.jsonEncode(),
         )
-    }
-
-    /**
-     * Throws an actionable error if the project with multiple Android variants has intersected source roots.
-     *
-     * See [org.jetbrains.dokka.gradle.engine.parameters.DokkaSourceSetSpec.basedOnAndroidVariant]
-     * for more information on why we need it.
-     */
-    private fun verifySourceRootsIntersectionsBasedOnAndroidVariants(
-        sourceSets: Set<DokkaSourceSetSpec>,
-        projectDir: File
-    ) {
-        // no check in case there is only single source-set
-        if (sourceSets.size <= 1) return
-
-        // check only source sets based on Android variants;
-        // everything else should be handled by the default checker in analysis
-        val sourceRootsIntersections = detectSourceRootsIntersections(
-            sourceSets.filter { it.basedOnAndroidVariant.getOrElse(false) }
-        )
-
-        if (sourceRootsIntersections.isEmpty()) return
-
-        val intersectedSourceSets = mutableSetOf<String>().apply {
-            sourceRootsIntersections.forEach { (s1Name, s2Name, _) ->
-                add(s1Name)
-                add(s2Name)
-            }
-        }
-
-        error(
-            """
-            |Dokka cannot generate documentation for Android projects with multiple enabled variants that have common source roots.
-            |Please suppress all variants except ONE via the following configuration:
-            |dokka {
-            |    dokkaSourceSets.configureEach {
-            |        suppress.set(name != VARIANT_NAME)
-            |    }
-            |}
-            |Where VARIANT_NAME could be one of the following: $intersectedSourceSets
-            |For more information regarding the reasoning behind this restriction: https://github.com/Kotlin/dokka/issues/4472
-            |Common source roots:
-            |${
-                sourceRootsIntersections.joinToString("\n") { (s1Name, s2Name, intersectedPaths) ->
-                    "  '$s1Name' and '$s2Name' have common source roots (relative to project directory):\n${
-                        intersectedPaths.joinToString("\n") {
-                            "    - ${it.relativeTo(projectDir).invariantSeparatorsPath}"
-                        }
-                    }"
-                }
-            }""".trimMargin()
-        )
-    }
-
-    private data class SourceRootIntersections(
-        val s1Name: String,
-        val s2Name: String,
-        val intersectedPaths: Set<File>,
-    )
-
-    // based on SourceRootIndependentChecker
-    private fun detectSourceRootsIntersections(sourceSets: List<DokkaSourceSetSpec>): List<SourceRootIntersections> {
-        fun Set<File>.normalize() = mapTo(mutableSetOf(), File::normalize)
-        fun intersectOfNormalizedPaths(normalizedPaths: Set<File>, normalizedPaths2: Set<File>): Set<File> {
-            val result = mutableSetOf<File>()
-            for (p1 in normalizedPaths) {
-                for (p2 in normalizedPaths2) {
-                    if (p1.startsWith(p2) || p2.startsWith(p1)) {
-                        result.add(p1)
-                        result.add(p2)
-                    }
-                }
-            }
-            return result
-        }
-
-        fun intersect(paths: Set<File>, paths2: Set<File>): Set<File> {
-            return intersectOfNormalizedPaths(paths.normalize(), paths2.normalize())
-        }
-
-        // no check in case there is only single source-set
-        if (sourceSets.size <= 1) return emptyList()
-
-        val intersections = mutableListOf<SourceRootIntersections>()
-        for (i in sourceSets.indices) {
-            for (j in i + 1 until sourceSets.size) {
-                val intersection = intersect(sourceSets[i].sourceRoots.files, sourceSets[j].sourceRoots.files)
-                if (intersection.isNotEmpty()) {
-                    intersections.add(
-                        SourceRootIntersections(
-                            s1Name = sourceSets[i].name,
-                            s2Name = sourceSets[j].name,
-                            intersectedPaths = intersection
-                        )
-                    )
-                }
-            }
-        }
-        return intersections
     }
 }
